Use row polymorphism for effect inference.
Functions are now by default effect-polymorphic rather than pure. This
is a relaxation in usage only, and doesn't weaken any effects. Effect
annotations are still required to talk about effects that a function
adds or removes.

There are now two effect kinds, `ERow` and `EScalar`, which behave much
like `Row` and `Scalar` but at the effect level. At present there is no
way to get a `Var` with kind `EScalar`, but that is not enforced.